+++ include/slicerecon/reconstruction/reconstructor.hpp
@@ -31,31 +31,12 @@ extern "C" {
 #include "../util/data_types.hpp"
 #include "../util/exceptions.hpp"
 #include "../util/log.hpp"
+#include "../util/processing.hpp"
 #include "helpers.hpp"
 
 namespace slicerecon {
 
 class reconstructor;
-
-struct paganin_settings {
-    float pixel_size;
-    float lambda;
-    float delta;
-    float beta;
-    float distance;
-};
-
-struct settings {
-    int32_t slice_size;
-    int32_t preview_size;
-    int32_t group_size;
-    int32_t filter_cores;
-    int32_t darks;
-    int32_t flats;
-    bool already_linear;
-    bool retrieve_phase;
-    paganin_settings paganin;
-};
 
 class listener {
   public:
@@ -304,21 +285,11 @@ class reconstructor {
     std::vector<float> small_volume_buffer_;
 
     std::vector<float> sino_buffer_;
-    std::vector<float> fdk_weights_;
-
-    bulk::thread::environment environment_;
 
     std::vector<listener*> listeners_;
     bool initialized_ = false;
 
-    fftwf_plan fft_plan_;
-    fftwf_plan ffti_plan_;
-    fftwf_plan fft2d_plan_;
-    fftwf_plan ffti2d_plan_;
-    std::vector<std::vector<std::complex<float>>> freq_buffer_;
-    std::vector<std::vector<std::complex<float>>> proj_freq_buffer_;
-    std::vector<float> filter_;
-    std::vector<float> paganin_filter_;
+    std::unique_ptr<util::ProjectionProcessor> projection_processor_;
 };
 
 } // namespace slicerecon