+++ CHANGELOG.md
@@ -11,11 +11,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `already_linear` flag, for already linearized data.
 - Add support for skipping flat fielding step.
 - Add Gaussian filter for optional noise suppression
+- Add support for Paganin filtering
 
 ### Changed
 - Change `plugin::listen` to run on the main thread
 - Change default value for flat field to `1`, darks and flats are now optional (#6).
 - Use FFTW3 as the FFT backend (instead of an unsupported Eigen module).
+- More modular system for processing projections
 
 ### Fixed
 - Fix application of FDK weighting.